[refactor] #31 refactor routine template add page layout
